+++ migrations/20240315155250_create_elem_rating_type.sql
@@ -1,0 +1,2 @@
+CREATE TYPE elem_rating AS (   elem text
+                            ,  rating real)