+++ src/core/Flora/Model/Package/Query.hs
@@ -6,13 +6,14 @@ module Flora.Model.Package.Query
   , countPackagesByName
   , countPackagesInNamespace
   , getAllPackageDependents
-  , getRequirementsQuery
   , getAllPackageDependentsWithLatestVersion
   , getAllPackages
   , getAllRequirements
   , getComponent
+  , getComponentById
   , getNonDeprecatedPackages
   , getNumberOfPackageDependents
+  , getNumberOfPackageRequirements
   , getPackageByNamespaceAndName
   , getPackageCategories
   , getPackageDependents
@@ -21,14 +22,15 @@ module Flora.Model.Package.Query
   , getPackagesByNamespace
   , getPackagesFromCategoryWithLatestVersion
   , getRequirements
+  , getRequirementsQuery
   , listAllPackages
   , listAllPackagesInNamespace
   , numberOfPackageRequirementsQuery
+  , searchExecutable
   , searchPackage
-  , unsafeGetComponent
-  , getComponentById
   , searchPackageByNamespace
-  , getNumberOfPackageRequirements
+  , unsafeGetComponent
+  , getNumberOfExecutablesByName
   ) where
 
 import Data.Text (Text)
@@ -65,6 +67,7 @@ import Flora.Model.Category.Types (PackageCategory)
 import Flora.Model.Component.Query qualified as Query
 import Flora.Model.Component.Types
 import Flora.Model.Package (Namespace (..), Package, PackageId, PackageInfo, PackageName (..))
+import Flora.Model.Package.Types (PackageInfoWithExecutables (..))
 import Flora.Model.Release.Types (ReleaseId)
 import Flora.Model.Requirement
   ( ComponentDependencies
@@ -541,11 +544,81 @@ searchPackageByNamespace (offset, limit) namespace searchString =
           , lv."version"
           , lv."license"
         ORDER BY rating desc, count(lv."namespace") desc, lv.name asc
-        OFFSET ?
         LIMIT ?
+        OFFSET ?
         ;
         |]
-      (searchString, searchString, namespace, offset, limit)
+      (searchString, searchString, namespace, limit, offset)
+
+searchExecutable
+  :: DB :> es
+  => (Word, Word)
+  -> Text
+  -> Eff es (Vector PackageInfoWithExecutables)
+searchExecutable (offset, limit) searchString =
+  dbtToEff $
+    query
+      Select
+      [sql|
+WITH results AS (SELECT DISTINCT l2.namespace
+                      , l2.name
+                      , l2.synopsis
+                      , l2.version
+                      , l2.license
+                      , word_similarity(p0.component_name, ?) AS rating
+                      , p0.component_name
+                 FROM package_components AS p0
+                      INNER JOIN releases AS r1 ON p0.release_id = r1.release_id
+                      INNER JOIN latest_versions AS l2 ON r1.package_id = l2.package_id
+                 WHERE p0.component_type = 'executable'
+                   AND ? <% p0.component_name)
+
+   , executables AS (SELECT DISTINCT r.namespace
+                      , r.name
+                      , r.synopsis
+                      , r.version
+                      , r.license
+                      , array_agg(((r.component_name, r.rating)::elem_rating) ORDER BY r.rating) AS execs
+                     FROM results AS r
+                     GROUP BY r.namespace, r.name, r.synopsis, r.version, r.license)
+
+  SELECT e.*
+  FROM executables AS e
+  ORDER BY (e.execs)[1].rating DESC
+
+LIMIT ?
+OFFSET ?
+        |]
+      (searchString, searchString, limit, offset)
+
+getNumberOfExecutablesByName :: DB :> es => Text -> Eff es Word
+getNumberOfExecutablesByName queryString = do
+  dbtToEff $ do
+    (result :: Maybe (Only Int)) <-
+      queryOne
+        Select
+        [sql|
+WITH results AS (SELECT l2.name
+                      , word_similarity(p0.component_name, ?) AS rating
+                      , p0.component_name
+                 FROM package_components AS p0
+                      INNER JOIN releases AS r1 ON p0.release_id = r1.release_id
+                      INNER JOIN latest_versions AS l2 ON r1.package_id = l2.package_id
+                 WHERE p0.component_type = 'executable'
+                   AND ? <% p0.component_name)
+
+   , executables AS (SELECT DISTINCT r.name
+                          , array_agg(CAST((r.component_name, r.rating) AS elem_rating) ORDER BY r.rating DESC) AS execs
+                     FROM results AS r
+                     GROUP BY r.name)
+
+  SELECT count(e.*)
+  FROM executables AS e
+          |]
+        (queryString, queryString)
+    case result of
+      Just (Only n) -> pure $ fromIntegral n
+      Nothing -> pure 0
 
 -- | Returns a summary of packages
 listAllPackages