+++ src/web/FloraWeb/Pages/Server/Search.hs
@@ -8,6 +8,7 @@ import Servant (Headers (..), ServerT)
 
 import Flora.Model.Package.Types
 import Flora.Model.User (User)
+import Flora.Search (SearchAction (..))
 import Flora.Search qualified as Search
 import FloraWeb.Common.Pagination
 import FloraWeb.Pages.Routes.Search (Routes, Routes' (..))
@@ -31,6 +32,39 @@ searchHandler (Headers session _) (Just searchString) pageParam = do
         templateDefaults
           { navbarSearchContent = Just searchString
           }
-  (count, results) <- Search.search (fromPage pageNumber) searchString
-  let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
-  render templateEnv $ Search.showResults searchString count pageNumber matchVector packagesInfo
+  let pagination = fromPage pageNumber
+  case Search.parseSearchQuery searchString of
+    Just (ListAllPackagesInNamespace namespace) -> do
+      (count, results) <- Search.listAllPackagesInNamespace pagination namespace
+      let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
+      render templateEnv $
+        Search.showResults searchString count pageNumber matchVector packagesInfo
+    Just ListAllPackages -> do
+      (count, results) <- Search.listAllPackages pagination
+      let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
+      render templateEnv $
+        Search.showResults searchString count pageNumber matchVector packagesInfo
+    Just (SearchInNamespace namespace (PackageName packageName)) -> do
+      (count, results) <- Search.searchPackageByNamespaceAndName pagination namespace packageName
+      let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
+      render templateEnv $
+        Search.showResults searchString count pageNumber matchVector packagesInfo
+    Just (DependentsOf namespace packageName mSearchString) -> do
+      (count, results) <- Search.searchDependents pagination namespace packageName mSearchString
+      let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
+      render templateEnv $
+        Search.showResults searchString count pageNumber matchVector packagesInfo
+    Just (SearchPackages _) -> do
+      (count, results) <- Search.searchPackageByName pagination searchString
+      let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
+      render templateEnv $
+        Search.showResults searchString count pageNumber matchVector packagesInfo
+    Nothing -> do
+      (count, results) <- Search.searchPackageByName pagination searchString
+      let (matchVector, packagesInfo) = Vector.partition (\p -> p.name == PackageName searchString) results
+      render templateEnv $
+        Search.showResults searchString count pageNumber matchVector packagesInfo
+    Just (SearchExecutable executableName) -> do
+      (count, results) <- Search.searchExecutable pagination executableName
+      render templateEnv $
+        Search.showExecutableResults searchString count pageNumber results