+++ src/web/FloraWeb/Pages/Templates/Screens/Search.hs
@@ -7,12 +7,12 @@ import Data.Text.Display (display)
 import Data.Vector (Vector)
 import Lucid
 
-import Flora.Model.Package (Namespace, PackageInfo (..))
+import Flora.Model.Package (Namespace, PackageInfo (..), PackageInfoWithExecutables (..))
 import Flora.Search (SearchAction (..))
 import FloraWeb.Components.PackageListHeader (presentationHeader)
 import FloraWeb.Components.PaginationNav (paginationNav)
 import FloraWeb.Pages.Templates
-import FloraWeb.Pages.Templates.Packages (packageListing)
+import FloraWeb.Pages.Templates.Packages (packageListing, packageWithExecutableListing)
 
 showAllPackages :: Word -> Positive Word -> Vector PackageInfo -> FloraHTML
 showAllPackages count currentPage packagesInfo = do
@@ -21,7 +21,13 @@ showAllPackages count currentPage packagesInfo = do
     div_ [class_ ""] $ packageListing Nothing packagesInfo
     paginationNav count currentPage ListAllPackages
 
-showAllPackagesInNamespace :: Namespace -> Text -> Word -> Positive Word -> Vector PackageInfo -> FloraHTML
+showAllPackagesInNamespace
+  :: Namespace
+  -> Text
+  -> Word
+  -> Positive Word
+  -> Vector PackageInfo
+  -> FloraHTML
 showAllPackagesInNamespace namespace description count currentPage packagesInfo = do
   div_ [class_ "container"] $ do
     presentationHeader (display namespace) description count
@@ -43,3 +49,17 @@ showResults searchString count currentPage exactMatches results = do
     packageListing (Just exactMatches) results
     when (count > 30) $
       paginationNav count currentPage (SearchPackages searchString)
+
+showExecutableResults
+  :: Text
+  -> Word
+  -> Positive Word
+  -> Vector PackageInfoWithExecutables
+  -- ^ Results
+  -> FloraHTML
+showExecutableResults executableName count currentPage results = do
+  div_ [class_ "container"] $ do
+    presentationHeader executableName "" count
+    packageWithExecutableListing results
+    when (count > 30) $
+      paginationNav count currentPage (SearchExecutable executableName)